Clean: move handle funcs from splitter.c to splitter_handlers.c
